\name{@ALIGNERINDEXNAME@}
\docType{package}

\alias{@PKGNAME@-package}
\alias{@PKGNAME@}
\alias{@ALIGNERINDEXNAME@}


\title{@PKGTITLE@}

\description{
  @PKGDESCRIPTION@
}

\details{
\tabular{ll}{
Package: \tab @PKGNAME@ \cr
Type: \tab Package\cr
Version: \tab @PKGVERSION@ \cr
Date: \tab @DATE@ \cr
License: \tab @LIC@ \cr
LazyLoad: \tab yes\cr
}
  @PKGDETAILS@
}

\note{
  This @ALIGNER@ index package was made from the following source data files:
  \preformatted{
@SRCDATAFILES@
  }
}

\author{@AUTHOR@}

\seealso{
}

\examples{
@PKGEXAMPLES@
}

\keyword{package}
\keyword{data}
